pyproject.toml
[project]
name = "document-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofile>=3.9.0",
    "githubkit>=0.13.0",
    "niquests>=3.14.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "uvloop>=0.21.0",
]

[tool.ruff.lint]
select = ["W", "E", "F", "I", "N", "Q", "A", "LOG", "PTH", "ANN", "PL", "RUF"]

[tool.uv]
package = false

[tool.mypy]
strict = true

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pre-commit>=4.2.0",
    "ruff>=0.12.7",
    "types-pyyaml>=6.0.12.20250516",
]
